Update DB slightRefactor-start
Update not pushed live at time of update due to a few bugs. Wanted to
get commit in.

Fixed typo with mist.utils.NMToMeters causing function not to work.
added Logger 'dbLog' to mist DB code
inserted do-end blocks to make DB code block shorter in length

Purpose of DB updating refactor is to more accurately keep unit tables
updated correctly. In the old system only 'byName' DBs were updated if a
group/unit name was duplicated. So it was possible for multiple units to
share the same name in unitsByNum, byCat, etc.

The process of how it checks added groups and units has been altered
slightly. Groups spawned with player controlled units will not be
updated.

The conditions for whether or not the table gets updated are as follows:

If groupName is already in DB then check if its groupId, country, or
coalition changed. If nothing has changed then it checks if the newly
spawned units exist in a DB entry and if so it will verify if the unitId
ort ype have changed. Any single update prompts the DB to update
accordingly.