Neutral coalition support
Added support for DB info in the miz regarding neutral groups. Removed the checks for the coalition name so that if there are ever more coalitions added to the game the mist code won't need to explicitly look for them.

Added local function to verify the DB entries for anything that was spawned before mist is loaded.

Set all popup messages to false.