Database Update Changes
ADDED: dbNum entry to unit table to define which index the the unit is at in the unitsByNum DB table
OPTIMIZED: multiple calls of getPosition to just be a single getPoint call
MODIFIED: process for updating and writing DB files. Moved the editing of DB into its own function
ADDED: error messages and handling if a DB entry fails to be updated
ADDED: mist.forceAddToDB function which accepts a groupName or a static objects name to force add it to the databases.