spawnsByBase and other fixes
ADDED: mist.DBs.spawnsByBase table that lists each airbase or unit that has an aircraft set to spawn on it.
MODIFIED: DB to contain the starting helipadId or airdromeId for each unit
MODIFIED: mist.getCurrentGroupData to use an empty table so it doesn't error if nothing is found
MODIFIED: mist.getCurrentGroupData to only call getPosition once.
MODIFIED: mist.getGroupTable to add country and category values to the table returned so it can be directly used in mist.dynAdd without having to dd those vlaue sback.